Move scoreboard/team specific components
