+++ src/navigation/root/tab/spirit/ScoreBoardScreen/Scoreboard/index.ts
