+++ src/navigation/root/tab/spirit/ScoreBoardScreen/ScoreboardItem/ScoreboardItem.tsx
