+++ src/navigation/root/tab/spirit/ScoreBoardScreen/ScoreboardItem/index.ts
