+++ src/navigation/root/tab/spirit/TeamScreen/TeamInformation/index.ts
